[TWM-153] Remove press sync from entrypoint.
We are moving the press db sync step to a k8 cronjob (sidecar solution
is more complicated in this instance).

This commit removes the press db sync fro the entrypoint and adds
a pressdb sync for local docker environment via `make db-sync`

This also add a little bit of cleanup for the Makefile to duplication
errors.