DEVO-544: Update default gems to secure versions.
This updates the docker file to update some gems included in the ruby
standard library that have security issues.

Updating these types of gems is tricker than non standard library gems
because ruby makes it harder to delete them.

Also, because they are part of the standard library we cannot make too
many assumptions about whether or not they are being used by our
application because normally any gem that we depend on could have
a dependency on one of these gems and only add a dependency on
a specific Ruby version (i.e. not the gem that it is using from the
standard library)

This is why I've only deleted the rdoc gem, since I can be reasonably
confident that we do not have a dependency on this gem. But, I cannot
make that assumption about the bundler, cgi, and rexml gems.