TWM-494: Help speed up pages.
Recently the TUPress front page has taken longer and longer to download.
I have determined that this is being caused because we are downloading
images from s3 to check for width and height even when the width and
height are already present.  This change updates the code to check that
image has not been "analzyed" before "analyzing" it which makes the
image get downloaded locally.

The s3 downloads are cumulative and vary in how long they they from 10 to
100s to 1000s of milliseconds each.